Do not swallow OSError when adding file watches
I found that tests were suddenly failing for no apparent reason. After investigation,
it turns out that vscode had exhausted my system's inotify quota, but we were eating
the relevant exception.

The exception handler in question was really only targeted at the "cannot open file" case,
which was already handled elsewhere, but was acting as a very delayed "remove the
file we can't read from our list of files" action.

I hate our PendingTasks system anyway, so I've made PendingFigure immediately handle
removing the StaticAsset from the Page if it can't open the referenced file,
and gotten rid of the now-pointless base class PendingTask.
